Background and bug fixes

+++ Source/RayTracer/Canvas.cpp
@@ -28,7 +28,7 @@ void Canvas::Clear(const color4_t& color)
 	std::fill(m_buffer.begin(), m_buffer.end(), rgba);
 }
 
-void Canvas::DrawPoint(const glm::vec2& point, const color4_t& color)
+void Canvas::DrawPoint(const glm::ivec2& point, const color4_t& color)
 {
 	if (point.x < 0 || point.x >= m_size.x || point.y < 0 || point.y >= m_size.y) return;
